Update import paths following moving of files
