+++ src/hooks/useChains.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getChains } from "~/api/chains";
+import { getChains } from "~/api/adamik/chains";
 
 export const useChains = () => {
   return useQuery({
