+++ src/hooks/useTransaction.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TransactionEncodeResponse } from "~/api/encode";
+import { TransactionEncodeResponse } from "~/api/adamik/encode";
 
 type TransactionContextType = {
   transaction: TransactionEncodeResponse | undefined;
