+++ src/hooks/useValidators.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getValidators } from "~/api/validator";
+import { getValidators } from "~/api/adamik/validators";
 
 type GetValidatorParams = {
   chainId: string;
