+++ src/providers/TransactionProvider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState } from "react";
-import { TransactionEncodeResponse } from "~/api/encode";
+import { TransactionEncodeResponse } from "~/api/adamik/encode";
 import { TransactionContext } from "~/hooks/useTransaction";
 
 export const TransactionProvider: React.FC<React.PropsWithChildren> = ({
